operator-flow/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.14
idna==3.4
importlib-metadata==6.0.1
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-zipkin-json==1.18.0
opentelemetry-exporter-zipkin-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
protobuf==3.20.3
requests==2.31.0
typing_extensions==4.6.3
urllib3==2.0.3
websockets==11.0.3
wrapt==1.15.0
zipp==3.15.0
